Move Function interfaction to internal package
