+++ viewcelladapter/src/main/java/ca/antonious/viewcelladapter/internal/Function.java
@@ -1,4 +1,4 @@
-package ca.antonious.viewcelladapter;
+package ca.antonious.viewcelladapter.internal;
 
 /**
  * Created by George on 2016-12-19.
